pyproject.toml
[project]
name = "watsonx-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Kiyeon Jeon",email = "kiyeon.jeon.21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "ibm-watsonx-ai>=1.4.2",
    "ibm-cos-sdk>=2.14.0",
    "boto3>=1.28.0",
    "chromadb>=0.4.0",
    "pymilvus>=2.3.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.4",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langchain-milvus>=0.1.0",
    "langchain-ibm>=0.3.0",
    "langgraph>=0.2.0",
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
